(S2) "Perfected" normal maps finally? Added debug attributes to materials
Turns out vNormalWs and uv tangent dont need to be used, as passing the normal to the shadingmodel handels it automatically?

v0 is now (0,0,1) instead of (1,1,1)
v1 is now (1,0,0) instead of i.vNormalWs
v2 is now (0,1,0) instead of uv tangent

This also helps with parallax stuff.
------
Added debug attributes to materials for use in the D2 Shader Baker tool (S&Box only)