Move everything Source 2 related to Source2Handler, create copy of ConfigHandler for Field
-Everything involving Source 2 has been moved to its own class (shader, vmat and vmdl exports)

-Created a copy of ConfigHandler to use in Field since it originally was Charm only
(Probably a good reason for this, but I did it anyways).
This makes getting shader export bools easier and wont make both UE and S2 shaders export if only UE was enabled, and vise versa

idk if it was always like this but the program needs a restart for config changes to actually take effect?